+++ oshi-core/src/main/java/oshi/hardware/platform/unix/dragonflybsd/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides information about hardware such as Memory, Power Sources, and Processor on DragonFly BSD systems
+ */
+package oshi.hardware.platform.unix.dragonflybsd;