+++ oshi-core/src/main/java/oshi/software/os/unix/dragonflybsd/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides information about Software and OS on DragonFly BSD
+ */
+package oshi.software.os.unix.dragonflybsd;