+++ oshi-core/src/main/java/oshi/util/platform/unix/dragonflybsd/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides utilities for DragonFly BSD
+ */
+package oshi.util.platform.unix.dragonflybsd;